pyproject.toml
[project]
name = "mcp-caldav"
version = "2.0.0"
description = "Multi-tenant MCP server for CalDAV and ICS calendar integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "caldav~=3.1.0",
    "mcp[cli]>=1.8.0",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.0.0",
    # Database (multi-tenant credential lookup)
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    # ICS feed parsing
    "icalendar>=6.0.0",
    "httpx>=0.27.0",
    # Redis (ICS feed cache + rate limiter backing store)
    "redis[hiredis]>=5.0.0",
    # Rate limiting (coredis is required by limits for async Redis storage)
    "limits>=4.0.0",
    "coredis>=4.12.0",
    # Credential encryption
    "cryptography>=43.0.0",
    # ASGI server
    "uvicorn>=0.30.0",
    "starlette>=0.40.0",
]

[project.scripts]
mcp-caldav = "mcp_caldav:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "uv>=0.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.4.0",
    "pre-commit>=3.6.0",
    "types-click>=7.1.0",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
pretty = true
error_summary = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "caldav.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mcp_caldav.server"
disable_error_code = ["misc"]

[tool.ruff]
target-version = "py311"
line-length = 88
extend-exclude = ["legacy"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused-function-args
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "Q", # flake8-quotes
    "RUF", # ruff-specific rules
]
ignore = [
    "E501", # line too long, handled by formatter
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "RUF002", # ambiguous unicode characters
    "RUF003", # ambiguous unicode characters in comments
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101", "PLR2004"]

[tool.ruff.lint.isort]
known-first-party = ["mcp_caldav"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
multiline-quotes = "double"
